fix float

+++ ft_ntoa_base.c
@@ -6,16 +6,16 @@
 /*   By: mleblanc <mleblanc@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/07 18:39:26 by mleblanc          #+#    #+#             */
-/*   Updated: 2021/06/15 20:40:13 by mleblanc         ###   ########.fr       */
+/*   Updated: 2021/06/17 22:54:53 by mleblanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_ntoa_base(char *b, long long n, size_t len, const char *base)
+void	ft_ntoa_base(char *b, long n, size_t len, const char *base)
 {
-	long long	base_len;
-	long long	i;
+	long	base_len;
+	long	i;
 
 	base_len = ft_strlen(base);
 	if (n < 0)
