+++ ft_numlen.c
@@ -6,13 +6,13 @@
 /*   By: mleblanc <mleblanc@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/07 18:21:02 by mleblanc          #+#    #+#             */
-/*   Updated: 2021/06/15 20:24:14 by mleblanc         ###   ########.fr       */
+/*   Updated: 2021/06/17 22:55:10 by mleblanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
-size_t	ft_numlen(long long n, long long radix)
+size_t	ft_numlen(long n, long radix)
 {
 	size_t	len;
